feat: Fd descriptor now Int32
Makes a lot of things so much easier. Also moved more intrinsics into
stdlib.

+++ crates/expo-codegen/src/builtins.rs
@@ -112,17 +112,17 @@ pub(crate) fn declare_builtins<'ctx>(
     // File I/O
     decl(
         "expo_fd_read",
-        ptr.fn_type(&[i64.into(), i64.into()], false),
+        ptr.fn_type(&[i32.into(), i64.into()], false),
     );
     decl(
         "expo_fd_write",
-        i64.fn_type(&[i64.into(), ptr.into(), i64.into()], false),
+        i64.fn_type(&[i32.into(), ptr.into(), i64.into()], false),
     );
-    decl("expo_fd_close", i64.fn_type(&[i64.into()], false));
+    decl("expo_fd_close", i32.fn_type(&[i32.into()], false));
     decl("expo_last_error", ptr.fn_type(&[], false));
     decl(
         "expo_file_open",
-        i64.fn_type(&[ptr.into(), i64.into()], false),
+        i32.fn_type(&[ptr.into(), i64.into()], false),
     );
     decl("expo_file_read_all", ptr.fn_type(&[ptr.into()], false));
     decl(
@@ -165,33 +165,33 @@ pub(crate) fn declare_builtins<'ctx>(
     );
 
     // Socket I/O
-    decl("expo_socket_create", i64.fn_type(&[i64.into()], false));
+    decl("expo_socket_create", i32.fn_type(&[i64.into()], false));
     decl(
         "expo_socket_bind",
-        i64.fn_type(&[i64.into(), ptr.into(), i64.into()], false),
+        i64.fn_type(&[i32.into(), ptr.into(), i64.into()], false),
     );
     decl(
         "expo_socket_connect",
-        i64.fn_type(&[i64.into(), ptr.into(), i64.into()], false),
+        i64.fn_type(&[i32.into(), ptr.into(), i64.into()], false),
     );
     decl("expo_socket_resolve", ptr.fn_type(&[ptr.into()], false));
     decl(
         "expo_socket_send_to",
-        i64.fn_type(&[i64.into(), ptr.into(), ptr.into(), i64.into()], false),
+        i64.fn_type(&[i32.into(), ptr.into(), ptr.into(), i64.into()], false),
     );
     decl(
         "expo_socket_recv_from",
-        ptr.fn_type(&[i64.into(), i64.into()], false),
+        ptr.fn_type(&[i32.into(), i64.into()], false),
     );
     decl(
         "expo_socket_listen",
-        i64.fn_type(&[i64.into(), i64.into()], false),
+        i64.fn_type(&[i32.into(), i64.into()], false),
     );
-    decl("expo_socket_accept", i64.fn_type(&[i64.into()], false));
-    decl("expo_socket_try_accept", i64.fn_type(&[i64.into()], false));
+    decl("expo_socket_accept", i32.fn_type(&[i32.into()], false));
+    decl("expo_socket_try_accept", i32.fn_type(&[i32.into()], false));
     decl(
         "expo_socket_setsockopt_reuse",
-        i64.fn_type(&[i64.into()], false),
+        i64.fn_type(&[i32.into()], false),
     );
 
     // Panic runtime
